refactor(convert): split postgres.rs into postgres/ module directory
Modularize the 3183-line postgres.rs into focused submodules for
maintainability. No logic, behavior, or API changes -- pure
restructuring with identical test results (120 tests pass).

New structure:
  postgres/
    mod.rs            -- config, struct, Backend trait impl
    correlation.rs    -- correlation/format query builder methods
    tests.rs          -- all unit tests
